fix: Refactor position handling to use double precision for coordinates across multiple classes

+++ src/main/java/com/devone/bot/core/brain/logic/navigator/finder/SimplePathUtils.java
@@ -13,20 +13,20 @@ public class SimplePathUtils {
     public static List<BotPosition> getSmartNeighbors(BotPosition current, Set<BotPosition> walkable) {
         List<BotPosition> result = new ArrayList<>();
     
-        int[][] directions = {
+        double[][] directions = {
             { 1, 0}, {-1, 0},
             { 0, 1}, { 0, -1}
         };
     
-        for (int[] d : directions) {
-            int dx = d[0];
-            int dz = d[1];
+        for (double[] d : directions) {
+            double dx = d[0];
+            double dz = d[1];
     
             // Пробуем шаги вверх, вниз и по уровню
-            for (int dy = -1; dy <= 1; dy++) {
-                int x = current.getX() + dx;
-                int y = current.getY() + dy;
-                int z = current.getZ() + dz;
+            for (double dy = -1; dy <= 1; dy++) {
+                double x = current.getX() + dx;
+                double y = current.getY() + dy;
+                double z = current.getZ() + dz;
     
                 BotPosition neighbor = new BotPosition(x, y, z);
